refactoring: remove MainPageController
- move methods from MainPageController to ConferencesController
- remove MainPageController class
- rename test class MainPageControllerTest -> ConferencesControllerTest
